requirements.txt
asgiref==3.2.10
Django==3.1.1
django-froala-editor==3.2.2
django-json-field==0.5.7
django-jsonfield==1.4.0
djangorestframework==3.11.1
pytz==2020.5
sqlparse==0.3.1
waitress==2.0.0
whitenoise==5.2.0
pymongo==3.11.4
requests==2.22.0
djangorestframework==3.11.1
sendgrid==6.4.8
celery==5.1.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
redis==3.5.3
cowin-vaccine-api==0.0.3